Move populate... methods to `AndroidEventProcessor` class
