+++ lib/src/main/kotlin/at/bitfire/ical4android/AndroidEvent.kt
@@ -20,58 +20,18 @@ import android.provider.CalendarContract.Events
 import android.provider.CalendarContract.EventsEntity
 import android.provider.CalendarContract.ExtendedProperties
 import android.provider.CalendarContract.Reminders
-import android.util.Patterns
 import androidx.core.content.contentValuesOf
 import at.bitfire.ical4android.AndroidEvent.Companion.CATEGORIES_SEPARATOR
 import at.bitfire.ical4android.AndroidEvent.Companion.numInstances
-import at.bitfire.ical4android.util.AndroidTimeUtils
-import at.bitfire.ical4android.util.DateUtils
 import at.bitfire.ical4android.util.MiscUtils.asSyncAdapter
-import at.bitfire.ical4android.util.TimeApiExtensions
-import at.bitfire.ical4android.util.TimeApiExtensions.toZonedDateTime
-import at.bitfire.synctools.exception.InvalidLocalResourceException
-import at.bitfire.synctools.icalendar.Css3Color
 import at.bitfire.synctools.mapping.calendar.AndroidEventBuilder
-import at.bitfire.synctools.mapping.calendar.AttendeeMappings
+import at.bitfire.synctools.mapping.calendar.AndroidEventProcessor
 import at.bitfire.synctools.storage.BatchOperation.CpoBuilder
 import at.bitfire.synctools.storage.LocalStorageException
 import at.bitfire.synctools.storage.calendar.AndroidCalendar
 import at.bitfire.synctools.storage.calendar.CalendarBatchOperation
-import at.bitfire.synctools.storage.removeBlank
-import at.bitfire.synctools.storage.toContentValues
-import net.fortuna.ical4j.model.Date
-import net.fortuna.ical4j.model.DateList
-import net.fortuna.ical4j.model.DateTime
-import net.fortuna.ical4j.model.component.VAlarm
-import net.fortuna.ical4j.model.parameter.Cn
-import net.fortuna.ical4j.model.parameter.Email
-import net.fortuna.ical4j.model.parameter.PartStat
-import net.fortuna.ical4j.model.parameter.Rsvp
-import net.fortuna.ical4j.model.parameter.Value
-import net.fortuna.ical4j.model.property.Action
-import net.fortuna.ical4j.model.property.Attendee
 import net.fortuna.ical4j.model.property.Clazz
-import net.fortuna.ical4j.model.property.Description
-import net.fortuna.ical4j.model.property.DtEnd
-import net.fortuna.ical4j.model.property.DtStart
-import net.fortuna.ical4j.model.property.ExDate
-import net.fortuna.ical4j.model.property.ExRule
-import net.fortuna.ical4j.model.property.Organizer
-import net.fortuna.ical4j.model.property.RDate
-import net.fortuna.ical4j.model.property.RRule
-import net.fortuna.ical4j.model.property.RecurrenceId
-import net.fortuna.ical4j.model.property.Status
-import net.fortuna.ical4j.model.property.Summary
-import net.fortuna.ical4j.util.TimeZones
 import java.io.FileNotFoundException
-import java.net.URI
-import java.net.URISyntaxException
-import java.time.Duration
-import java.time.Instant
-import java.time.Period
-import java.time.ZoneOffset
-import java.time.ZonedDateTime
-import java.util.logging.Level
 import java.util.logging.Logger
 
 /**
@@ -160,28 +120,13 @@ class AndroidEvent(
                 )
 
                 if (iterEvents.hasNext()) {
-                    val e = iterEvents.next()
+                    val entity = iterEvents.next()
+                    return Event().also { newEvent ->
+                        val processor = AndroidEventProcessor(calendar, id, entity)
+                        processor.populate(to = newEvent)
 
-                    // create new Event which will be populated
-                    val newEvent = Event()
-                    _event = newEvent
-
-                    // calculate some scheduling properties
-                    val groupScheduled = e.subValues.any { it.uri == Attendees.CONTENT_URI }
-
-                    populateEvent(e.entityValues.removeBlank(), groupScheduled)
-
-                    for (subValue in e.subValues) {
-                        val subValues = subValue.values.removeBlank()
-                        when (subValue.uri) {
-                            Attendees.CONTENT_URI -> populateAttendee(subValues)
-                            Reminders.CONTENT_URI -> populateReminder(subValues)
-                            ExtendedProperties.CONTENT_URI -> populateExtended(subValues)
-                        }
+                        _event = newEvent
                     }
-                    populateExceptions()
-                    useRetainedClassification()
-                    return newEvent
                 }
             } catch (e: Exception) {
                 /* Populating event has been interrupted by an exception, so we reset the event to
@@ -192,365 +137,9 @@ class AndroidEvent(
             } finally {
                 iterEvents?.close()
             }
+
             throw FileNotFoundException("Couldn't find event $id")
         }
-
-    /**
-     * Reads event data from the calendar provider.
-     *
-     * @param row values of an [Events] row, as returned by the calendar provider
-     */
-    private fun populateEvent(row: ContentValues, groupScheduled: Boolean) {
-        logger.log(Level.FINE, "Read event entity from calender provider", row)
-        val event = requireNotNull(event)
-
-        row.getAsString(Events.MUTATORS)?.let { strPackages ->
-            val packages = strPackages.split(MUTATORS_SEPARATOR).toSet()
-            event.userAgents.addAll(packages)
-        }
-
-        val allDay = (row.getAsInteger(Events.ALL_DAY) ?: 0) != 0
-        val tsStart = row.getAsLong(Events.DTSTART) ?: throw InvalidLocalResourceException("Found event without DTSTART")
-
-        var tsEnd = row.getAsLong(Events.DTEND)
-        var duration =   // only use DURATION of DTEND is not defined
-                if (tsEnd == null)
-                    row.getAsString(Events.DURATION)?.let { AndroidTimeUtils.parseDuration(it) }
-                else
-                    null
-
-        if (allDay) {
-            event.dtStart = DtStart(Date(tsStart))
-
-            // Android events MUST have duration or dtend [https://developer.android.com/reference/android/provider/CalendarContract.Events#operations].
-            // Assume 1 day if missing (should never occur, but occurs).
-            if (tsEnd == null && duration == null)
-                duration = Duration.ofDays(1)
-
-            if (duration != null) {
-                // Some servers have problems with DURATION, so we always generate DTEND.
-                val startDate = ZonedDateTime.ofInstant(Instant.ofEpochMilli(tsStart), ZoneOffset.UTC).toLocalDate()
-                if (duration is Duration)
-                    duration = Period.ofDays(duration.toDays().toInt())
-                tsEnd = (startDate + duration).toEpochDay() * TimeApiExtensions.MILLIS_PER_DAY
-                duration = null
-            }
-
-            if (tsEnd != null) {
-                when {
-                    tsEnd < tsStart ->
-                        logger.warning("dtEnd $tsEnd (allDay) < dtStart $tsStart (allDay), ignoring")
-
-                    tsEnd == tsStart ->
-                        logger.fine("dtEnd $tsEnd (allDay) = dtStart, won't generate DTEND property")
-
-                    else /* tsEnd > tsStart */ ->
-                        event.dtEnd = DtEnd(Date(tsEnd))
-                }
-            }
-
-        } else /* !allDay */ {
-            // use DATE-TIME values
-
-            // check time zone ID (calendar apps may insert no or an invalid ID)
-            val startTzId = DateUtils.findAndroidTimezoneID(row.getAsString(Events.EVENT_TIMEZONE))
-            val startTz = DateUtils.ical4jTimeZone(startTzId)
-            val dtStartDateTime = DateTime(tsStart).apply {
-                if (startTz != null) {  // null if there was not ical4j time zone for startTzId, which should not happen, but technically may happen
-                    if (TimeZones.isUtc(startTz))
-                        isUtc = true
-                    else
-                        timeZone = startTz
-                }
-            }
-            event.dtStart = DtStart(dtStartDateTime)
-
-            // Android events MUST have duration or dtend [https://developer.android.com/reference/android/provider/CalendarContract.Events#operations].
-            // Assume 1 hour if missing (should never occur, but occurs).
-            if (tsEnd == null && duration == null)
-                duration = Duration.ofHours(1)
-
-            if (duration != null) {
-                // Some servers have problems with DURATION, so we always generate DTEND.
-                val zonedStart = dtStartDateTime.toZonedDateTime()
-                tsEnd = (zonedStart + duration).toInstant().toEpochMilli()
-                duration = null
-            }
-
-            if (tsEnd != null) {
-                if (tsEnd < tsStart)
-                    logger.warning("dtEnd $tsEnd < dtStart $tsStart, ignoring")
-                /*else if (tsEnd == tsStart)    // iCloud sends 404 when it receives an iCalendar with DTSTART but without DTEND
-                    logger.fine("dtEnd $tsEnd == dtStart, won't generate DTEND property")*/
-                else /* tsEnd > tsStart */ {
-                    val endTz = row.getAsString(Events.EVENT_END_TIMEZONE)?.let { tzId ->
-                        DateUtils.ical4jTimeZone(tzId)
-                    } ?: startTz
-                    event.dtEnd = DtEnd(DateTime(tsEnd).apply {
-                        if (endTz != null) {
-                            if (TimeZones.isUtc(endTz))
-                                isUtc = true
-                            else
-                                timeZone = endTz
-                        }
-                    })
-                }
-            }
-
-        }
-
-        // recurrence
-        try {
-            row.getAsString(Events.RRULE)?.let { rulesStr ->
-                for (rule in rulesStr.split(AndroidTimeUtils.RECURRENCE_RULE_SEPARATOR))
-                    event.rRules += RRule(rule)
-            }
-            row.getAsString(Events.RDATE)?.let { datesStr ->
-                val rDate = AndroidTimeUtils.androidStringToRecurrenceSet(datesStr, allDay, tsStart) { RDate(it) }
-                event.rDates += rDate
-            }
-
-            row.getAsString(Events.EXRULE)?.let { rulesStr ->
-                for (rule in rulesStr.split(AndroidTimeUtils.RECURRENCE_RULE_SEPARATOR))
-                    event.exRules += ExRule(null, rule)
-            }
-            row.getAsString(Events.EXDATE)?.let { datesStr ->
-                val exDate = AndroidTimeUtils.androidStringToRecurrenceSet(datesStr, allDay) { ExDate(it) }
-                event.exDates += exDate
-            }
-        } catch (e: Exception) {
-            logger.log(Level.WARNING, "Couldn't parse recurrence rules, ignoring", e)
-        }
-
-        event.uid = row.getAsString(Events.UID_2445)
-        event.sequence = row.getAsInteger(COLUMN_SEQUENCE)
-        event.isOrganizer = row.getAsBoolean(Events.IS_ORGANIZER)
-
-        event.summary = row.getAsString(Events.TITLE)
-        event.location = row.getAsString(Events.EVENT_LOCATION)
-        event.description = row.getAsString(Events.DESCRIPTION)
-
-        // color can be specified as RGB value and/or as index key (CSS3 color of AndroidCalendar)
-        event.color =
-            row.getAsString(Events.EVENT_COLOR_KEY)?.let { name ->      // try color key first
-                try {
-                    Css3Color.valueOf(name)
-                } catch (_: IllegalArgumentException) {
-                    logger.warning("Ignoring unknown color name \"$name\"")
-                    null
-                }
-            } ?:
-            row.getAsInteger(Events.EVENT_COLOR)?.let { color ->        // otherwise, try to find the color name from the value
-                Css3Color.entries.firstOrNull { it.argb == color }
-            }
-
-        // status
-        when (row.getAsInteger(Events.STATUS)) {
-            Events.STATUS_CONFIRMED -> event.status = Status.VEVENT_CONFIRMED
-            Events.STATUS_TENTATIVE -> event.status = Status.VEVENT_TENTATIVE
-            Events.STATUS_CANCELED -> event.status = Status.VEVENT_CANCELLED
-        }
-
-        // availability
-        event.opaque = row.getAsInteger(Events.AVAILABILITY) != Events.AVAILABILITY_FREE
-
-        // scheduling
-        if (groupScheduled) {
-            // ORGANIZER must only be set for group-scheduled events (= events with attendees)
-            if (row.containsKey(Events.ORGANIZER))
-                try {
-                    event.organizer = Organizer(URI("mailto", row.getAsString(Events.ORGANIZER), null))
-                } catch (e: URISyntaxException) {
-                    logger.log(Level.WARNING, "Error when creating ORGANIZER mailto URI, ignoring", e)
-                }
-        }
-
-        // classification
-        when (row.getAsInteger(Events.ACCESS_LEVEL)) {
-            Events.ACCESS_PUBLIC -> event.classification = Clazz.PUBLIC
-            Events.ACCESS_PRIVATE -> event.classification = Clazz.PRIVATE
-            Events.ACCESS_CONFIDENTIAL -> event.classification = Clazz.CONFIDENTIAL
-        }
-
-        // exceptions from recurring events
-        row.getAsLong(Events.ORIGINAL_INSTANCE_TIME)?.let { originalInstanceTime ->
-            val originalAllDay = (row.getAsInteger(Events.ORIGINAL_ALL_DAY) ?: 0) != 0
-            val originalDate =
-                    if (originalAllDay)
-                        Date(originalInstanceTime)
-                    else
-                        DateTime(originalInstanceTime)
-            if (originalDate is DateTime) {
-                event.dtStart?.let { dtStart ->
-                    if (dtStart.isUtc)
-                        originalDate.isUtc = true
-                    else if (dtStart.timeZone != null)
-                        originalDate.timeZone = dtStart.timeZone
-                }
-            }
-            event.recurrenceId = RecurrenceId(originalDate)
-        }
-    }
-
-    private fun populateAttendee(row: ContentValues) {
-        logger.log(Level.FINE, "Read event attendee from calender provider", row)
-
-        try {
-            val attendee: Attendee
-            val email = row.getAsString(Attendees.ATTENDEE_EMAIL)
-            val idNS = row.getAsString(Attendees.ATTENDEE_ID_NAMESPACE)
-            val id = row.getAsString(Attendees.ATTENDEE_IDENTITY)
-
-            if (idNS != null || id != null) {
-                // attendee identified by namespace and ID
-                attendee = Attendee(URI(idNS, id, null))
-                email?.let { attendee.parameters.add(Email(it)) }
-            } else
-                // attendee identified by email address
-                attendee = Attendee(URI("mailto", email, null))
-            val params = attendee.parameters
-
-            // always add RSVP (offer attendees to accept/decline)
-            params.add(Rsvp.TRUE)
-
-            row.getAsString(Attendees.ATTENDEE_NAME)?.let { cn -> params.add(Cn(cn)) }
-
-            // type/relation mapping is complex and thus outsourced to AttendeeMappings
-            AttendeeMappings.androidToICalendar(row, attendee)
-
-            // status
-            when (row.getAsInteger(Attendees.ATTENDEE_STATUS)) {
-                Attendees.ATTENDEE_STATUS_INVITED -> params.add(PartStat.NEEDS_ACTION)
-                Attendees.ATTENDEE_STATUS_ACCEPTED -> params.add(PartStat.ACCEPTED)
-                Attendees.ATTENDEE_STATUS_DECLINED -> params.add(PartStat.DECLINED)
-                Attendees.ATTENDEE_STATUS_TENTATIVE -> params.add(PartStat.TENTATIVE)
-                Attendees.ATTENDEE_STATUS_NONE -> { /* no information, don't add PARTSTAT */ }
-            }
-
-            event!!.attendees.add(attendee)
-        } catch (e: URISyntaxException) {
-            logger.log(Level.WARNING, "Couldn't parse attendee information, ignoring", e)
-        }
-    }
-
-    private fun populateReminder(row: ContentValues) {
-        logger.log(Level.FINE, "Read event reminder from calender provider", row)
-        val event = requireNotNull(event)
-
-        val alarm = VAlarm(Duration.ofMinutes(-row.getAsLong(Reminders.MINUTES)))
-
-        val props = alarm.properties
-        when (row.getAsInteger(Reminders.METHOD)) {
-            Reminders.METHOD_EMAIL -> {
-                val accountName = calendar.account.name
-                if (Patterns.EMAIL_ADDRESS.matcher(accountName).matches()) {
-                    props += Action.EMAIL
-                    // ACTION:EMAIL requires SUMMARY, DESCRIPTION, ATTENDEE
-                    props += Summary(event.summary)
-                    props += Description(event.description ?: event.summary)
-                    // Android doesn't allow to save email reminder recipients, so we always use the
-                    // account name (should be account owner's email address)
-                    props += Attendee(URI("mailto", calendar.account.name, null))
-                } else {
-                    logger.warning("Account name is not an email address; changing EMAIL reminder to DISPLAY")
-                    props += Action.DISPLAY
-                    props += Description(event.summary)
-                }
-            }
-
-            // default: set ACTION:DISPLAY (requires DESCRIPTION)
-            else -> {
-                props += Action.DISPLAY
-                props += Description(event.summary)
-            }
-        }
-        event.alarms += alarm
-    }
-
-    private fun populateExtended(row: ContentValues) {
-        val name = row.getAsString(ExtendedProperties.NAME)
-        val rawValue = row.getAsString(ExtendedProperties.VALUE)
-        logger.log(Level.FINE, "Read extended property from calender provider", arrayOf(name, rawValue))
-        val event = requireNotNull(event)
-
-        try {
-            when (name) {
-                EXTNAME_CATEGORIES ->
-                    event.categories += rawValue.split(CATEGORIES_SEPARATOR)
-
-                EXTNAME_URL ->
-                    try {
-                        event.url = URI(rawValue)
-                    } catch(_: URISyntaxException) {
-                        logger.warning("Won't process invalid local URL: $rawValue")
-                    }
-
-                EXTNAME_ICAL_UID ->
-                    // only consider iCalUid when there's no uid
-                    if (event.uid == null)
-                        event.uid = rawValue
-
-                UnknownProperty.CONTENT_ITEM_TYPE ->
-                    event.unknownProperties += UnknownProperty.fromJsonString(rawValue)
-            }
-        } catch (e: Exception) {
-            logger.log(Level.WARNING, "Couldn't parse extended property", e)
-        }
-    }
-
-    private fun populateExceptions() {
-        requireNotNull(id)
-        val event = requireNotNull(event)
-
-        calendar.client.query(Events.CONTENT_URI.asSyncAdapter(calendar.account),
-                null,
-                Events.ORIGINAL_ID + "=?", arrayOf(id.toString()), null)?.use { c ->
-            while (c.moveToNext()) {
-                val values = c.toContentValues()
-                try {
-                    val exception = AndroidEvent(calendar, values)
-                    val exceptionEvent = exception.event!!
-                    val recurrenceId = exceptionEvent.recurrenceId!!
-
-                    // generate EXDATE instead of RECURRENCE-ID exceptions for cancelled instances
-                    if (exceptionEvent.status == Status.VEVENT_CANCELLED) {
-                        val list = DateList(
-                                if (DateUtils.isDate(recurrenceId)) Value.DATE else Value.DATE_TIME,
-                                recurrenceId.timeZone
-                        )
-                        list.add(recurrenceId.date)
-                        event.exDates += ExDate(list).apply {
-                            if (DateUtils.isDateTime(recurrenceId)) {
-                                if (recurrenceId.isUtc)
-                                    setUtc(true)
-                                else
-                                    timeZone = recurrenceId.timeZone
-                            }
-                        }
-
-                    } else /* exceptionEvent.status != Status.VEVENT_CANCELLED */ {
-                        // make sure that all components have the same ORGANIZER [RFC 6638 3.1]
-                        exceptionEvent.organizer = event.organizer
-
-                        // add exception to list of exceptions
-                        event.exceptions += exceptionEvent
-                    }
-                } catch (e: Exception) {
-                    logger.log(Level.WARNING, "Couldn't find exception details", e)
-                }
-            }
-        }
-    }
-
-    private fun retainClassification() {
-        /* retain classification other than PUBLIC and PRIVATE as unknown property so
-           that it can be reused when "server default" is selected */
-        val event = requireNotNull(event)
-        event.classification?.let {
-            if (it != Clazz.PUBLIC && it != Clazz.PRIVATE)
-                event.unknownProperties += it
-        }
-    }
 
 
     /**
